+++ cli/src/main/kotlin/sift/cli/SerializationOptions.kt
@@ -8,20 +8,14 @@ import java.io.File
 
 class SerializationOptions : OptionGroup(name = "Serialization options") {
     val save: File? by option("-s", "--save",
-            metavar = "FILE_JSON",
-            help = "Save the resulting system model as json.",
-            completionCandidates = CompletionCandidates.Path
+        metavar = "FILE_JSON",
+        help = "Save the resulting system model as json.",
+        completionCandidates = CompletionCandidates.Path
     ).file(canBeDir = false)
 
     val load: File? by option("--load",
-            metavar = "FILE_JSON",
-            help = "Load a previously saved system model.",
-            completionCandidates = CompletionCandidates.Path
-    ).file(canBeDir = false, mustExist = true, mustBeReadable = true)
-
-    val diff: File? by option("-d", "--diff",
-            metavar = "FILE_JSON",
-            help = "Compare against a previously saved system model.",
-            completionCandidates = CompletionCandidates.Path
+        metavar = "FILE_JSON",
+        help = "Load a previously saved system model.",
+        completionCandidates = CompletionCandidates.Path
     ).file(canBeDir = false, mustExist = true, mustBeReadable = true)
 }