+++ cli/src/main/kotlin/sift/cli/TemplateOptions.kt
@@ -6,6 +6,8 @@ import com.github.ajalt.clikt.parameters.options.convert
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import org.objectweb.asm.tree.ClassNode
+import sift.core.api.SystemModel
+import sift.core.api.resolveSystemModel
 import sift.core.asm.resolveClassNodes
 import sift.core.template.SystemModelTemplate
 
@@ -26,10 +28,15 @@ class TemplateOptions : OptionGroup(name = "Template options") {
         .flag()
 
     val classNodes: List<ClassNode>? by option("-f", "--class-dir", "--classes",
-            metavar = "PATH|URI",
-            help = "Path to directory structure containing classes or path to 'jar' file. " +
-                   "If the path is a URI, it must point to a 'jar' file.")
+            metavar = "PATH|URI|MAVEN_COORD",
+            help = "Provide class input as a directory, JAR file, URI (pointing to a JAR), or Maven coordinate.")
         .convert { resolveClassNodes(it) }
+
+    val diff: SystemModel? by option("-d", "--diff",
+            metavar = "FILE_JSON|URI|MAVEN_COORD",
+            help = "Compare the system model from '-f' with another, specified as a JSON file (previously saved System Model), " +
+                   "class input as a directory, JAR file, URI (pointing to a JAR), or Maven coordinate.")
+        .convert { resolveSystemModel(it, template) }
 
     val profile: Boolean by option("--profile",
             help = "Print execution times and input/output for the executed template.")